% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{add_haplotype_rainbow_axis_theming}
\alias{add_haplotype_rainbow_axis_theming}
\title{Add naming for the x-axis targets and reshape the output plot to have less empty space around plot}
\usage{
add_haplotype_rainbow_axis_theming(
  rainbow_plot,
  prepped_allele_data,
  sampleCol = s_Sample,
  targetCol = p_name
)
}
\arguments{
\item{rainbow_plot}{The rainbow plot object created the gen rainbow plot functions}

\item{prepped_allele_data}{the "preped" table created by the prep functions}

\item{sampleCol}{the sample name column to use, this does not need to be the same as what was used for gen rainbow plot but can be a new name but must be a factor with same number of levels}

\item{targetCol}{the character name of the target column to use for the naming of the columns, this does not need to be the same as what was used for gen rainbow plot but can be a new name but must be a factor with same number of levels}
}
\value{
the ggplot object
}
\description{
Add naming for the x-axis targets and reshape the output plot to have less empty space around plot
}
