% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{color_palettes}
\alias{color_palettes}
\alias{colorPalette_08}
\alias{colorPalette_12}
\alias{colorPalette_15}
\title{Colour-blind-friendly categorical palettes}
\format{
Character vectors of hex colour strings:
\describe{
\item{colorPalette_08}{8 colours.}
\item{colorPalette_12}{12 colours.}
\item{colorPalette_15}{15 colours.}
}
}
\usage{
colorPalette_08

colorPalette_12

colorPalette_15
}
\description{
Three categorical palettes (8, 12 and 15 colours) drawn from colour-blind-friendly
colour sets, suitable for colouring sample/target metadata bands or for use as the
\code{colors} ramp in \link{HaplotypeRainbow}'s \code{plot()}.
}
\examples{
scales::show_col(colorPalette_12)
}
