% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{genRainbowHapPlotObjActualFracLogColor}
\alias{genRainbowHapPlotObjActualFracLogColor}
\title{Create a ggplot object for the plotting by a rainblow of colors}
\usage{
genRainbowHapPlotObjActualFracLogColor(
  prepData,
  sampleCol = s_Sample,
  targetCol = p_name,
  popUIDCol = h_popUID,
  relAbundCol = c_AveragedFrac,
  colorCol = popidFracLogColor,
  colors = RColorBrewer::brewer.pal(11, "Spectral")
)
}
\arguments{
\item{prepData}{the prepped data, the columns used should be the same as the prepped}

\item{sampleCol}{the name of the sample column}

\item{targetCol}{the name of the target column}

\item{popUIDCol}{the name of the identifier column}

\item{relAbundCol}{the name of the relative abundance column}

\item{colorCol}{the color by which to color}

\item{colors}{the colors to spread across, the number supplied should be the same number given to the prep functions}
}
\value{
a ggplot2 object that can be further modified
}
\description{
Create a ggplot object for the plotting by a rainblow of colors
}
