% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{genRainbowHapPlotObjShade}
\alias{genRainbowHapPlotObjShade}
\title{Create a ggplot object for the plotting by a generating shades of input colors}
\usage{
genRainbowHapPlotObjShade(
  prepData,
  sampleCol = s_Sample,
  targetCol = p_name,
  popUIDCol = h_popUID,
  relAbundCol = c_AveragedFrac,
  colorCol = h_color_byFreq_mod
)
}
\arguments{
\item{prepData}{the prepped data, the columns used should be the same as the prepped}

\item{sampleCol}{the name of the sample column}

\item{targetCol}{the name of the target column}

\item{popUIDCol}{the name of the identifier column}

\item{relAbundCol}{the name of the relative abundance column}

\item{colorCol}{the color by which to color}
}
\value{
a ggplot2 object that can be further modified
}
\description{
Create a ggplot object for the plotting by a generating shades of input colors
}
