% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HaplotypeRainbows-package.R
\docType{package}
\name{HaplotypeRainbows-package}
\alias{HaplotypeRainbows}
\alias{HaplotypeRainbows-package}
\title{HaplotypeRainbows: Creates Haplotype Rainbow Plots}
\description{
Creates haplotype "rainbow" plots from targeted amplicon (microhaplotype) sequencing data. Given a table of samples, targets, within-target haplotype identifiers and relative counts, it builds a grid where columns are targets and rows are samples, with each cell showing the within-sample haplotype composition. Haplotype colours rotate across targets to produce the characteristic rainbow. The package exposes a single R6 class, HaplotypeRainbow, that carries the column mapping so it only has to be set once. Column defaults follow the Portable Microhaplotype Object (PMO) convention.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nickjhathaway/HaplotypeRainbows}
  \item Report bugs at \url{https://github.com/nickjhathaway/HaplotypeRainbows/issues}
}

}
\author{
\strong{Maintainer}: Nicholas Hathaway \email{nickjhathaway@gmail.com} (\href{https://orcid.org/0000-0001-9639-2894}{ORCID})

Authors:
\itemize{
  \item Nicholas Hathaway \email{nickjhathaway@gmail.com} (\href{https://orcid.org/0000-0001-9639-2894}{ORCID})
}

}
\keyword{internal}
