% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{prepForRainbow}
\alias{prepForRainbow}
\title{Prep a dataframe for clustering functions, this sorts the alleles by their population rank
The input data is processed so that the information needed for the haplotype rainbow plotting functions}
\usage{
prepForRainbow(
  inputData,
  sampleCol = s_Sample,
  targetCol = p_name,
  popUIDCol = h_popUID,
  relAbundCol = c_AveragedFrac,
  minPopSize = 1,
  colorOuput = 11,
  barHeight = 0.8
)
}
\arguments{
\item{inputData}{the input data}

\item{sampleCol}{the name of the sample column}

\item{targetCol}{the name of the target column}

\item{popUIDCol}{the name of the identifier column}

\item{relAbundCol}{the name of the relative abundance column}

\item{minPopSize}{the minimum population size of a target, if a target has less than unique alleles than the target is dropped}

\item{colorOuput}{the number of colors to spread across}

\item{barHeight}{the height of the final full bars per sample, controls whether they touch of not, e.g. barHeight==1 will create touching bars}
}
\value{
the prepped data frame for the plotting function
}
\description{
Prep a dataframe for clustering functions, this sorts the alleles by their population rank
The input data is processed so that the information needed for the haplotype rainbow plotting functions
}
