% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{prepForRainbowShade}
\alias{prepForRainbowShade}
\title{Prep a dataframe for clustering functions, this preps for coloring by shading rather than by a rainbow of colors
The input data is processed so that the information needed for the haplotype rainbow plotting functions}
\usage{
prepForRainbowShade(
  inputData,
  sampleCol = s_Sample,
  targetCol = p_name,
  popUIDCol = h_popUID,
  relAbundCol = c_AveragedFrac,
  minPopSize = 3,
  baseColors = c("#e41a1c", "#377eb8", "#4daf4a", "#984ea3", "#ff7f00", "#ffff33"),
  barHeight = 0.8
)
}
\arguments{
\item{inputData}{the input data}

\item{sampleCol}{the name of the sample column}

\item{targetCol}{the name of the target column}

\item{popUIDCol}{the name of the identifier column}

\item{relAbundCol}{the name of the relative abundance column}

\item{minPopSize}{the minimum population size of a target, if a target has less than unique alleles than the target is dropped}

\item{baseColors}{the colors to create shades from}

\item{barHeight}{the height of the final full bars per sample, controls whether they touch of not, e.g. barHeight==1 will create touching bars}
}
\value{
the prepped data frame for the plotting function
}
\description{
Prep a dataframe for clustering functions, this preps for coloring by shading rather than by a rainbow of colors
The input data is processed so that the information needed for the haplotype rainbow plotting functions
}
