% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFunctions.R
\name{resort_prepped_samples_by_clustering}
\alias{resort_prepped_samples_by_clustering}
\title{resort_prepped_samples_by_clustering
Reset the sample factor levels to be in order that similar samples based on haplotype sharing}
\usage{
resort_prepped_samples_by_clustering(
  prepped_allele_data,
  sampleCol = s_Sample,
  targetCol = p_name,
  popUIDCol = h_popUID,
  relAbundCol = c_AveragedFrac,
  target_sample_coverage_freq_cut_off = 0.8,
  by_major_allele = FALSE
)
}
\arguments{
\item{prepped_allele_data}{the prepped data table}

\item{sampleCol}{the name of the sample column}

\item{targetCol}{the name of the target column}

\item{popUIDCol}{the name of the identifier column}

\item{relAbundCol}{the name of the relative abundance column}

\item{target_sample_coverage_freq_cut_off}{the sample coverage per target cut off, targets must have at least this freq of sample coverage to be used in the clustering}

\item{by_major_allele}{whether or not to cluster just by the major allele}
}
\value{
the prepped data frame with the resorted sample levels
}
\description{
resort_prepped_samples_by_clustering
Reset the sample factor levels to be in order that similar samples based on haplotype sharing
}
